+++ CHANGELOG.md
@@ -4,6 +4,8 @@ NOWnews-api change logs
 
 ## [Unreleased]
 
+- 軍聞社匯入新聞改分類 @appleoxxo
+
 ## 1.0.39 - 2017-10-17
 ### Changed
 - 後台上傳圖片先去打一次imagelab協助併版 @wayne